Fix the size for the disk quota info
In quota_maps, it stores both size and limit in it.
In function calculate_table_disk_usage, we only use
the "increment" part to update the size in quota_map.

So, if an entry in quota_map is vacuumed last time
and its size is some value, such as xx, but as it
has been vacuumed, then in the next loop its size
will be 0 or 0 + increment
Or for a new disk quota, as we don't init size for
it, its initial value is always 0, this is not correct
either.

So when we create a new entry in quota_map, we need
firstly init the size for it.

And at the same time, I think there is no need to vacuum
the quota info entry when the limit is -1, as we may use
it later. It may lead the quota map bigger and bigger,
but I think it's ok, as there arent' a lot of schemas/
roles/tablespaces in a database.

Link the test about tablespaces into upgrade foler.