Support index in a different tablespace with the original table
* Use pg_table_size instead of pg_total_realtion_size
to get the disk usage. pg_table_size includes all forks,
toast, and toast index, except the index.

* Add a function to get the list oids of the tables which
the diskquota needs to care about in the database.

firstly , get the tables which rekind is “r” or “m” and
not system tables
 "select oid from pg_class"
 " where oid >= %u and (relkind='r' or relkind='m')",
 FirstNormalObjectId);
then fetch the indexes of those tables and add them to the
list.

* Use the list of oids in init_table_size_table and calculate_table_disk_usage functions