Fix CRATE TABLE when diskquota is not preloaded
Currently, tables created when diskquota is not in the
shared_preload_libraries will not be counted into quota usage even
after the user calls diskquota.init_table_size_table(). This is because
the user have no way to tell the bgworker that it needs to reload
the table sizes.

The fix addes a flag in the worker map to indicate that the worker
needs to reload the table sizes. Each time the user calls init(), the
flag is set so that no table update will be missed during the time
when diskquota is disabled (i.e., not preloaded).