Fix empty quota view when no table
Currently, quota views are implemented using INNER JOIN. As a result,
when there is no table belong to a role, schema, or tablespace, the
quota view is empty even though there is a quota config defined for
that role, schema, or tablespace.

This patch fixes this by re-implementing quota views using OUTER JOIN.
Plus, the implementation also uses CTEs for modularizing big queries
and aggregation pushdown.