Pack the last released so in the pipeline
Due to our upgrade design, the newly released diskquota should contain
all the so file of every latest minor release. The packaging system has
been implemented by cmake DISKQUOTA_PREVIOUS_INSTALLER.
This commit get the latest release through gcs regexp and pass it to the
build script.

The pipeline has been spited into different jobs based on the OS type
because of:
- Easier to use the common resource name for build_diskquota.yml
- job_def.lib.yml becomes shorter
- Can span jobs into different concourse workers, which may bring some
  better performance.
- Running flaky test again is easier.

Also a gate job has been added which is the main entrance for all
pipelines. In the future, we may want to have different gate job for
different pipelines, like a clang-format checking gate for the PR
pipeline.