Fix change tablespace test case
"ALTER TABLE xx SET TABLESPACE" will change the relfilenode of
a table, it will copy files from old relfilenode directory to
the new one and delete all files when commiting the transaction.

As diskquota doesn't acquire a lock on the relation when
fetching table size, so if diskquota is collecting active
tables' size and another session is deleting a active table's
files under a old relfilenode directory in a changing
tablespace transaction, the diskquota may can't get all segment's
table size as files on part of segments have been deleted.

To fix it, we try to make a table to be an active table by
inserting data into it and wait the diskquota to recalculate
the tablesize after changing tablespace.