Fix bugs and refactor
Bug fix:
- Move epoch from DiskQuotaWorkerEntry to DiskquotaDBEntry
- reset the diskquota worker shmem by reset_disk_quota_model
when dropping the extension
- Move the database list from laucher local memory to
shared memory, so other process can use: such the diskquota
bg worker can set database status.

Refactor:
remove the map disk_quota_worker_map, as it is a mess.

It mainly records if a diskquota extension paused and stores
the bgworker's handler. And it should be distributed to
segments as the segment hooks need to know if a database is
paused when hardlimit is "on".

As there is map "monitoring_dbid_cache" used to recorded which
database are monitored by diskquota, and it needs to be
dispatched to segments too, so I just use it to implement the
fuction of pausing a database. If a database paused, just
removing it from "monitoring_dbid_cache", and add it back when
resuming. "monitoring_dbid_cache" is used as a white list.

There is a side affect of the refactor, when a database is
paused, diskquota will not collect the active tables of it,
add after resuming it, the table size of it can not be updated
immediately, but I don't think it's a big deal.