Log readiness message only once
Currently a diskquota worker will log an error message when the
database is not ready to be mornitored by Diskquota once for each
iteration. When the naptime is small, the log will be flooded by this
error message.

This patch adds a flag in the worker map in the shared  memory so
that the error message will only be logged once per startup.