Close free gang in launcher and worker
As launcher and worker both run in bgworker, they are clients when
connecting to segments. They know when should to close the connections.
So we directly call `DisconnectAndDestroyAllGangs` to close them, and
remove the code that only colses free read gang based on timeout.

For launcher, we destroy gangs after init_datbase_list or after each
time processing create/drop diskquota extension.

For worker, we only destroy gangs after the extension is paused.