Fix test failure on release pipeline
Regress tests failed on release
pipeline due to the diskquota.naptime
is not 0. As diskquota launcher
firstly sleep, then starts diskquota
workers. When naptime is not 0, workers
are started very late, then some tests
which checking them failed.

In this pr, fix the test by starting
diskquota workers in launcher before
it sleeps. This makes diskquota workers
will be started immediately when launcher
starts.

Also, increased the sleep time in
the tests.