Store all configs in a row for each database.
Put all configs into a `config jsonb` and set a version field for this
config.  When `JSON_parse_quota_config` finds the version field in
`config jsonb` is the lower version, it will call the `do_parser`
with the corresponding version. And `JSON_construct_quota_config` will
call the `do_constr` with the current version, and fill the version
field as the current version.