Quota config version 3.0
Store all configs in a row for each database.
- Put all configs into a `config jsonb` and set a version field for this
config.  When `JSON_parse_quota_config` finds the version field in
`config jsonb` is the lower version, it will call the `do_parser`
with the corresponding version. And `JSON_construct_quota_config` will
call the `do_constr` with the current version, and fill the version
field as the current version.

Merge quota_config 3.0 with old code.
- Add diskquota--3.0.sql
- Move some tool functions to diskquota_util.c
- Modify do_load_quotas(), use quota_config 3.0 to load quota configs
  and update quota_map.
- Update version to 3.0
- Workaround: ignore diskquota.quota_config for is_database_empty().
Ignore the toast aux table and its index of diskquota.quota_config.

Some modification for regress tests.
- Add diskquota.show_quota_config_view
- Modify fast quota view for diskquota.quota_config 3.0
- Add super user check for role quota, we should not set role quota for
  super user.
- Add a judgement for `segratio == 0`, 0 is a invalid segratio for
  tablespace quota.
- The update for quota_config does not support concurrency, so we do
  not test in parallel.
- test_default_tablespace is disabled.
- test_per_segment_config is disabled.
- test_rejectmap_mul_db is disabled for some bugs.
- test_tablespace_schema_perseg is updated.
- Disable upgrade test