Improve the performance of refresh_rejectmap().
Previously, in refresh_rejectmap(), we remove all entries in disk_quota_reject_map
and then insert new entries into it. In this produce, some entries haved been removed
and inserted, which waste too much time to occupy the reject_map_lock.

This PR adds a hashtable `local_duplicate_rejectmap` to record which entries do not need
to be removed. We just insert/remove entries that do not exist in `local_duplicate_rejectmap`,
so that we can reduce the time for occupying the reject_map_lock.