Fix diskquota worker schedule bug
When a database's diskquota bgworker is a killed and the db is droped,
diskquota scheduler can not work properly. The cause is: if the scheduler
failed to start a bgworker for a database, it will try it again and
again forever.

Fixing it by: firstly check if the database is still valid, if not, skip
it in the next_db routine. Secondly, limit the times of starting a
bgworker for a database, if has tried 3 times, will skip it and pick the
next one.