Fix bug: gpstart timeout for gpdb7.
We used to set `Gp_role = GP_ROLE_DISPATCH` in
disk_quota_launcher_main(), even though postmaster boots in utility
mode. This seems to be nothing in gpdb6, but it will cause a dead
loop when booting gpdb7. In fact, there is nothing to do in utility
mode for the diskquota launcher. In this commit, if
`Gp_role != GP_ROLE_DISPATCH`, disk_quota_launcher_main() will
simply exit.