Quota config 3.0 (#306)
## Problem
The old version of the quota config management is not friendly to upgrade/downgrade.
For example, if a new type of quota config needs to be added, the table structure where
the quota config is stored must be modified. This makes it difficult for diskquota to
transfer quota config data from a lower version to a higher version(or from a higher
version to a lower version).

Meanwhile, the quota config is stored in both `diskquota.quota_config` and `diskquota.target`,
which increases the cost of data maintenance.

## Design
This PR refactors quota config relevant tables in diskquota: `diskquota.target` is removed
and `diskquota.quota_config` is refactored. Quota configs are stored as JSON in
`diskquota.quota_config`, for example:
```
{
    "version": "diskquota-3.0",
    "quota_list": [
        {
            "quota_type": 0,            //NAMESPACE_QUOTA
            "db_oid": 123,
            "namespace_oid": 1,
            "quota_limit_mb": 100
        },
        {
            "quota_type": 1,          //ROLE_QUOTA
            "db_oid": 123,
            "owner_oid": 2,
            "quota_limit_mb": 100
        }
    ]
}
```
The quota config update procedure is:
- load JSON string of all quota configs from `diskquota.quota_config` to local memory.
- parse the JSON string to quota_config structs by JSON parse function, and store them 
  in `quota_config_map`(a hash table).
- update quota config in `quota_config_map`.
- construct JSON string by `quota_config_map`.
- store JSON string into `diskquota.quota_config`.

In the future version, if a new type of quota config needs to be added, we just need to 
update the JSON parse function, instead of changing the structure of the
`diskquota.quota_config` table.

Update version to 3.0:
- Update diskquota--3.0.sql
  - Add diskquota.show_quota_config_view
  - Modify fast quota view for quota_config 3.0
- Move some tool functions from diskquota_utility.c to diskquota_util.c
- Modify do_load_quotas(), use quota_config 3.0 to load quota configs and update quota_map.
- Workaround: ignore diskquota.quota_config for is_database_empty(). Ignore the toast
  aux table and its index of diskquota.quota_config.

Some modifications for regress tests.
- The update for quota_config does not support concurrency, so we do not test in parallel.
- test_tablespace_schema_perseg is updated.
- Upgrade test is disabled.
- test_rejectmap_mul_db is disabled for flaky caused by some bugs.
- test_default_tablespace is disabled. Default role and database(tablespace) exceed quota,
  which interrupts the update for diskquota.quota_config.
- test_per_segment_config is disabled. The update for quota_config does not support concurrency.