Free cache tuple after use
refresh_rejectmap looked for an oid in local cache and retrieved a copy
in a case of success. However, such a tuple should be freed after use,
but refresh_rejectmap didn't do that which resulted in huge memory leak
if there were passed lots of oid, which could result in diskquota
stopping working. This patch frees these tuples and prevets memory leaks